refactor #77: hooks 디렉터리 구조 수정
